requirements.txt
 transformers
gradio
clip-by-openai
Pillow
numpy
tqdm